fix problem with expanded selection and focusing to beginning of input
There was problem for all browsers that when user expanded selection (CMD+A) and hit left arrow key, it'd move focus to beginning of input. This was problem when tags were already present as you could start typing. Now I enforce typing query only at the end of input. This should also prevent problems on Firefox.