add CSS rule for `@avocode/styled-avocode-email-tagsinput` to override document `user-select` property
If document, where the input is rendered includes `-webkit-user-select:
none` the SlateJS library will throw error and crash when trying to
programatically add tags to the input. The email input won't
receive any input which is fine but the error is hard to debug and does
not lead to problems regarding this CSS property.

Since this library is opinionated, I decided to include this CSS rule
with it.